backend/microservices/auth/requirements.txt
isort
black
fastapi
asyncpg
psycopg2
alembic
ormar
uvicorn
PyJWT
python-dotenv
validate_email